Move applyPatchToSnapshot back into revisions/util
The leaf module existed only to break a cycle in the old design where the
adapter imported the event-dispatch service. Now that dispatch lives in the
handler-layer webhook-adapter registry (not the adapter), toApiSavedGroupRevision
-> util -> index -> adapter no longer loops back, so the extraction is no longer
needed. Verified no init cycle.
