+++ packages/back-end/src/revisions/util.ts
@@ -1,4 +1,6 @@
-import { isEqual } from "lodash";
+import { applyPatch } from "fast-json-patch";
+import type { Operation } from "fast-json-patch";
+import { cloneDeep, isEqual } from "lodash";
 import {
   JsonPatchOperation,
   Revision,
@@ -8,12 +10,26 @@ import {
 import { ReqContext } from "back-end/types/request";
 import { ApiReqContext } from "back-end/types/api";
 import { getAdapter } from "back-end/src/revisions/index";
-// Re-exported from the leaf module so existing `revisions/util` importers keep
-// working, while serializers (toApiSavedGroupRevision) import it directly from
-// the leaf to avoid an import cycle through the adapter registry.
-import { applyPatchToSnapshot } from "back-end/src/revisions/applyPatch";
 
-export { applyPatchToSnapshot };
+/**
+ * Apply a set of JSON Patch ops to a snapshot, returning a new object.
+ *
+ * Clones with lodash `cloneDeep` (which preserves Date instances) and lets
+ * applyPatch mutate that throwaway copy in place (mutateDocument = true).
+ * Passing mutateDocument = false would make fast-json-patch internally
+ * JSON-clone the input, converting Date fields (e.g. dateCreated/dateUpdated on
+ * a saved-group snapshot) into ISO strings and breaking downstream serializers
+ * that call `.toISOString()`.
+ */
+export function applyPatchToSnapshot<T extends object>(
+  snapshot: T,
+  proposedChanges: JsonPatchOperation[] | unknown,
+): T {
+  const ops = normalizeProposedChanges(proposedChanges);
+  if (ops.length === 0) return snapshot;
+  return applyPatch(cloneDeep(snapshot), ops as Operation[], false, true)
+    .newDocument as T;
+}
 
 /**
  * Ensure a "live" merged revision exists representing the entity's current state.
